// src/app/share/enums/colors.enum.ts
export enum ColorMapTitle {
  black = 'Black',
  blue = 'Blue',
  red = 'Red',
  white = 'White',
  green = 'Green',
  silver = 'Silver',
  spaceGrey = 'Space Gray',
  midnight = 'Midnight',
  starlight = 'Starlink',
  gold = 'Gold',
  pink = 'Pink',
  purple = 'Purple',
}
export enum ColorMap {
  black = 1,
  blue = 2,
  green = 3,
  red = 4,
  white = 5,
  silver = 6,
  spaceGrey = 7,
  midnight = 8,
  starlight = 9,
  gold = 10,
  pink = 11,
  purple = 12,
}
export const ParsColorsToString: any = {
  [ColorMap.black]: 'Black',
  [ColorMap.blue]: 'Blue',
  [ColorMap.green]: 'Green',
  [ColorMap.red]: 'Red',
  [ColorMap.white]: 'White',
  [ColorMap.silver]: 'Silver',
  [ColorMap.spaceGrey]: 'Space Grey',
  [ColorMap.midnight]: 'Midnight',
  [ColorMap.starlight]: 'Starlight',
  [ColorMap.gold]: 'Gold',
  [ColorMap.pink]: 'Pink',
  [ColorMap.purple]: 'Purple',
};